remove files in warnings/ from foobar/CMakeLists.txt

+++ src/foobar/ports/vasprintf.c
@@ -60,7 +60,7 @@ int vasprintf(char **buf, const char *fmt, va_list ap)
     } /* CAW: old glibc versions have this problem */
 #endif /* defined(WIN32) */
 
-    b = (char *)malloc(sizeof(char) * chars);
+    b = (char *)malloc(sizeof(char) * (size_t)chars);
     if (!b) {
         return -1;
     }
